pyproject.toml
[project]
name = "vakilsahab-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "dill>=0.4.1",
    "dotenv>=0.9.9",
    "keybert>=0.9.0",
    "langchain>=1.2.12",
    "langchain-aws>=1.4.0",
    "langchain-chroma>=1.1.0",
    "langchain-huggingface>=1.2.1",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=1.1.1",
    "numpy>=2.4.3",
    "opendatasets>=0.1.22",
    "pandas>=3.0.1",
    "pyyaml>=6.0.3",
    "langchain-community",
    "sentence-transformers>=5.3.0",
    "fastmcp",
    "fastapi",
    "uvicorn[standard]",
    "fastapi-mcp>=0.4.0",
    "langchain-mcp-adapters",
]
